auth file is working as intended
