+++ src/auth/index.js
@@ -37,3 +37,27 @@ authenticate = (jwt, next) => {
     };
  };
  
+ export const signout = (next) => {
+    if(typeof window !== "undefined") localStorage.removeItem("jwt") //delete the local storage contents of the jwt &
+                next(); // then make sure we log them out of our backend api
+    return fetch("http://localhost:8080/signout", {
+    method: "GET"
+})
+    .then( response  => {
+          console.log('signout', response)
+          return response.json()
+    })
+    .catch(err => console.log(err))
+}
+
+export const isAuthenticated = () => {
+    if(typeof window == "undefined") {
+          return false
+    }
+    if(localStorage.getItem("jwt")) {
+          return JSON.parse(localStorage.getItem("jwt"))
+    } else {
+          return false
+    }
+};
+
